back/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
pytz==2022.2.1
pywatchman==1.4.1
sqlparse==0.4.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-templated-mail==1.1.1
gunicorn==20.1.0
djoser==2.1.0